+++ packages/react-examples/package.json
@@ -66,6 +66,7 @@
     "@fluentui/react-image": "^0.6.19-4",
     "@fluentui/react-internal": "^8.0.0-beta.7",
     "@fluentui/react-link": "^1.0.0-beta.7",
+    "@fluentui/react-shared-contexts": "^1.0.0-beta.0",
     "@fluentui/react-slider": "^1.0.0-beta.7",
     "@fluentui/react-tabs": "^1.0.0-beta.7",
     "@fluentui/react-text": "^0.1.3-4",
