+++ packages/react-examples/src/react-button/MenuButton/MenuButton.stories.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
-import { useMenuContext, MenuButton, MenuButtonProps, MinimalMenuProps } from '@fluentui/react-button';
+import { MenuButton, MenuButtonProps } from '@fluentui/react-button';
 import { Callout, ContextualMenu, IContextualMenuProps, Stack, Text } from '@fluentui/react';
+import { useMenuContext, MinimalMenuProps } from '@fluentui/react-shared-contexts';
 import * as classes from '../Button.stories.scss';
 
 const menuProps = {
