+++ packages/react-shared-contexts/.eslintrc.json
@@ -1,0 +1,7 @@
+{
+  "extends": ["plugin:@fluentui/eslint-plugin/react--legacy"],
+  "root": true,
+  "rules": {
+    "prefer-const": "off"
+  }
+}
