+++ packages/react-shared-contexts/.npmrc
@@ -1,0 +1,2 @@
+registry=https://registry.npmjs.org/
+
