+++ packages/react-shared-contexts/README.md
@@ -1,0 +1,8 @@
+# @fluentui/react-shared-contexts
+
+**Shared contexts for [Fluent UI React](https://developer.microsoft.com/en-us/fluentui)**
+([formerly Office UI Fabric React](https://developer.microsoft.com/en-us/office/blogs/ui-fabric-is-evolving-into-fluent-ui/))
+
+This package includes a number of exported React Contexts that are shared by Fluent UI components.
+
+See [GitHub](https://github.com/microsoft/fluentui) for more details on the Fluent UI React project and packages within.
