+++ packages/react-shared-contexts/config/api-extractor.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@fluentui/scripts/api-extractor/api-extractor.common.json"
+}
