+++ packages/react-shared-contexts/etc/react-shared-contexts.api.md
@@ -1,0 +1,25 @@
+## API Report File for "@fluentui/react-shared-contexts"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import * as React from 'react';
+
+// @public (undocumented)
+export const MenuContext: React.Context<MinimalMenuProps>;
+
+// @public (undocumented)
+export type MinimalMenuProps = {
+    hidden?: boolean;
+    onDismiss?: () => void;
+    target?: React.Ref<HTMLElement | undefined>;
+};
+
+// @public (undocumented)
+export const useMenuContext: () => MinimalMenuProps;
+
+
+// (No @packageDocumentation comment for this package)
+
+```
