+++ packages/react-shared-contexts/just.config.ts
@@ -1,0 +1,3 @@
+import { preset } from '@fluentui/scripts';
+
+preset();
