+++ packages/react-shared-contexts/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@fluentui/react-internal",
-  "version": "8.0.0-beta.7",
-  "description": "Component internals for @fluentui/react.",
+  "name": "@fluentui/react-shared-contexts",
+  "version": "1.0.0-beta.0",
+  "description": "Fluent UI React Contexts shared by multiple components.",
   "main": "lib-commonjs/index.js",
   "module": "lib/index.js",
   "sideEffects": [
@@ -15,39 +15,32 @@
   "license": "MIT",
   "scripts": {
     "build": "just-scripts build",
-    "clean": "just-scripts clean",
-    "code-style": "just-scripts code-style",
-    "just": "just-scripts",
+    "bundle": "just-scripts bundle",
     "lint": "just-scripts lint",
-    "start-test": "just-scripts jest-watch",
     "test": "just-scripts test",
+    "just": "just-scripts",
+    "code-style": "just-scripts code-style",
+    "clean": "just-scripts clean",
+    "start-test": "just-scripts jest-watch",
     "update-api": "just-scripts update-api",
     "update-snapshots": "just-scripts jest -u"
   },
   "devDependencies": {
     "@fluentui/eslint-plugin": "^1.0.0-beta.0",
-    "@types/chalk": "^2.0.0",
     "@types/enzyme": "3.10.3",
     "@types/enzyme-adapter-react-16": "1.0.3",
-    "@types/glob": "^7.1.1",
     "@types/jest": "~24.9.0",
     "@types/node": "^10.3.2",
+    "@types/prop-types": "15.7.1",
     "@types/react": "16.8.25",
     "@types/react-dom": "16.8.4",
     "@types/react-test-renderer": "^16.0.0",
     "@types/sinon": "2.2.2",
-    "@types/webpack-env": "1.15.1",
     "@fluentui/scripts": "^1.0.0",
-    "@fluentui/example-data": "^8.0.0-beta.0",
     "@fluentui/jest-serializer-merge-styles": "^8.0.0-beta.1",
     "@fluentui/test-utilities": "^8.0.0-beta.1",
-    "@fluentui/react-conformance": "^1.0.0",
-    "chalk": "^2.1.0",
     "enzyme": "~3.10.0",
     "enzyme-adapter-react-16": "^1.15.0",
-    "glob": "^7.1.2",
-    "jest-snapshot": "~24.9.0",
-    "office-ui-fabric-core": "^11.0.0",
     "react": "16.8.6",
     "react-app-polyfill": "~1.0.1",
     "react-dom": "16.8.6",
@@ -55,26 +48,12 @@
     "sinon": "^4.1.3"
   },
   "dependencies": {
-    "@fluentui/date-time-utilities": "^8.0.0-beta.0",
-    "@fluentui/foundation-legacy": "^8.0.0-beta.5",
-    "@fluentui/font-icons-mdl2": "^8.0.0-beta.5",
-    "@fluentui/merge-styles": "^8.0.0-beta.1",
-    "@fluentui/react-focus": "^8.0.0-beta.5",
-    "@fluentui/react-hooks": "^8.0.0-beta.5",
-    "@fluentui/react-shared-contexts": "^1.0.0-beta.0",
-    "@fluentui/react-theme-provider": "^1.0.0-beta.6",
-    "@fluentui/react-window-provider": "^2.0.0-beta.1",
     "@fluentui/set-version": "^8.0.0-beta.0",
-    "@fluentui/style-utilities": "^8.0.0-beta.5",
-    "@fluentui/theme": "^2.0.0-beta.5",
-    "@fluentui/utilities": "^8.0.0-beta.5",
-    "@microsoft/load-themed-styles": "^1.10.26",
+    "prop-types": "^15.7.2",
     "tslib": "^1.10.0"
   },
   "peerDependencies": {
     "@types/react": ">=16.8.0 <17.0.0",
-    "@types/react-dom": ">=16.8.0 <17.0.0",
-    "react": ">=16.8.0 <17.0.0",
-    "react-dom": ">=16.8.0 <17.0.0"
+    "react": ">=16.8.0 <17.0.0"
   }
 }
