+++ packages/react-shared-contexts/src/MenuContext.ts
@@ -1,0 +1,3 @@
+export * from './MenuContext/index';
+
+import './version';
