+++ packages/react-shared-contexts/src/MenuContext/MenuContext.ts
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { MinimalMenuProps } from './types';
+
+export const MenuContext = React.createContext<MinimalMenuProps>({});
+
+export const useMenuContext = () => {
+  return React.useContext(MenuContext);
+};
