+++ packages/react-shared-contexts/src/MenuContext/types.ts
@@ -1,0 +1,10 @@
+import * as React from 'react';
+
+/**
+ * {@docCategory MenuContext}
+ */
+export type MinimalMenuProps = {
+  hidden?: boolean;
+  onDismiss?: () => void;
+  target?: React.Ref<HTMLElement | undefined>;
+};
