+++ packages/react-shared-contexts/src/index.ts
@@ -1,0 +1,3 @@
+export * from './MenuContext';
+
+import './version';
